requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
line_profiler==4.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.3.6
pluggy==0.13.1
py==1.11.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.2
requests==2.26.0
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
starlette==0.41.3
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.1
